Tuned some coefficients for the game.

+++ space-shooter/Assets/Assets/Scripts/Lifetime.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Lifetime : MonoBehaviour {
 
-    public int lifetimeInSeconds;
+    public float lifetimeInSeconds;
 
 	// Use this for initialization
 	void Start () {
